requirements/local.txt
-r production.txt
Werkzeug[watchdog]==3.1.3 # https://github.com/pallets/werkzeug
ipdb==0.13.13  # https://github.com/gotcha/ipdb
watchfiles==1.1.0  # https://github.com/samuelcolvin/watchfiles

# Testing
# ------------------------------------------------------------------------------
mypy==1.18.2  # https://github.com/python/mypy
django-stubs[compatible-mypy]==5.2.6  # https://github.com/typeddjango/django-stubs
pytest==8.4.2  # https://github.com/pytest-dev/pytest
pytest-sugar==1.1.1  # https://github.com/Teemu/pytest-sugar
pytest-cov==7.0.0
djangorestframework-stubs==3.16.4  # https://github.com/typeddjango/djangorestframework-stubs
debugpy==1.8.17
boto3-stubs[essential]==1.42.7
mypy-boto3-dynamodb==1.42.3


# Documentation
# ------------------------------------------------------------------------------
sphinx==8.2.3 # pyup: != 8.3.0  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2025.8.25 # https://github.com/GaretJax/sphinx-autobuild
pydotplus==2.0.2

# Code quality
# ------------------------------------------------------------------------------
ruff==0.13.3  # https://github.com/astral-sh/ruff
coverage==7.10.7  # https://github.com/nedbat/coveragepy
djlint==1.36.4  # https://github.com/Riverside-Healthcare/djLint
pre-commit==4.3.0  # https://github.com/pre-commit/pre-commit
pylint==3.3.9
pylint_django==2.6.1
types_pytz==2025.2.0.20251108
types-python-dateutil==2.9.0.20251115

# Django
# ------------------------------------------------------------------------------
factory-boy==3.3.2  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==6.0.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==4.1  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==3.2.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.11.1  # https://github.com/pytest-dev/pytest-django
